added tests for mongodb connection
